Cleanup RecordTransactionCommit()
Remove a log message to note if a QE reader tried to commit.  This used
to happen back when lazy XID didn't exist and a reader started a
transaction.  StartTransaction() used to generate a new XID and if the
new XID resulted in CLOG extension, the QE reader would end up writing
an XLOG record.  This case can no longer arise with lazy XID.

Also cleanup a few unused variables nearby.